Image=wagrun:dev, separate Database connection env vars
It looked like using dj_database_url wasn't parsing the env var
property, handing Postgres a DB Name > 73 chars (the length of the
entire url?) and its max is 63. So I split them out and now it's
parsing those fine.

But we're timing out going to aurora -- is it not waking up, or is the
connectivity dead? Maybe we should not scale-to-zero yet.